add post api.

+++ internal/dto/post_dto.go
@@ -2,5 +2,5 @@ package dto
 
 type PostDto struct {
 	Content  string `json:"content"`
-	AuthorID uint   `json:"authorId"`
+	AuthorID uint64 `json:"authorId"`
 }
